chore: migrate to Gradle 9 and update dependencies
