+++ androidApp/src/main/AndroidManifest.xml
@@ -18,7 +18,7 @@
         android:supportsRtl="true"
         android:theme="@android:style/Theme.Material.Light.NoActionBar">
         <activity
-            android:name=".MainActivity"
+            android:name="com.ma7moud3ly.quran.MainActivity"
             android:exported="true">
             <intent-filter>
                 <action android:name="android.intent.action.MAIN" />
@@ -27,7 +27,7 @@
             </intent-filter>
         </activity>
         <service
-            android:name=".PlaybackService"
+            android:name="com.ma7moud3ly.quran.PlaybackService"
             android:exported="false"
             android:foregroundServiceType="mediaPlayback" />
     </application>
