+++ composeApp/build.gradle.kts
@@ -7,26 +7,25 @@ import java.util.Properties
 
 plugins {
     alias(libs.plugins.kotlinMultiplatform)
-    alias(libs.plugins.androidApplication)
+    alias(libs.plugins.androidKmpLibrary)
     alias(libs.plugins.composeMultiplatform)
     alias(libs.plugins.composeCompiler)
-    alias(libs.plugins.composeHotReload)
     alias(libs.plugins.kotlinSerialization)
     alias(libs.plugins.ksp)
 }
 
-// apply gms & firebase plugin only for gms build flavour
-if (gradle.startParameter.taskRequests.toString().contains("gms", ignoreCase = true)) {
-    apply(plugin = libs.plugins.google.services.get().pluginId)
-    apply(plugin = libs.plugins.firebase.crashlytics.get().pluginId)
-}
-
-
 kotlin {
-    androidTarget {
-        @OptIn(ExperimentalKotlinGradlePluginApi::class)
+    androidLibrary {
+        namespace = libs.versions.project.packageName.get()
+        compileSdk = libs.versions.android.compileSdk.get().toInt()
+        minSdk = libs.versions.android.minSdk.get().toInt()
+
         compilerOptions {
             jvmTarget.set(JvmTarget.JVM_11)
+        }
+
+        androidResources {
+            enable = true
         }
     }
 
@@ -47,16 +46,12 @@ kotlin {
     wasmJs {
         outputModuleName.set("composeApp")
         browser {
-            val rootDirPath = project.rootDir.path
-            val projectDirPath = project.projectDir.path
             commonWebpackConfig {
                 outputFileName = "composeApp.js"
                 devServer = (devServer ?: KotlinWebpackConfig.DevServer()).apply {
-                    static = (static ?: mutableListOf()).apply {
-                        // Serve sources to debug inside browser
-                        add(rootDirPath)
-                        add(projectDirPath)
-                    }
+                    // Serve sources to debug inside browser
+                    static(directory = project.rootDir.path, false)
+                    static(directory = project.projectDir.path, false)
                 }
             }
         }
@@ -65,13 +60,13 @@ kotlin {
 
     sourceSets {
         commonMain.dependencies {
-            implementation(compose.runtime)
-            implementation(compose.foundation)
-            implementation(compose.material3)
-            implementation(compose.materialIconsExtended)
-            implementation(compose.ui)
-            implementation(compose.components.resources)
-            implementation(compose.components.uiToolingPreview)
+            implementation(libs.compose.runtime)
+            implementation(libs.compose.foundation)
+            implementation(libs.compose.material3)
+            implementation(libs.compose.materialIconsExtended)
+            implementation(libs.compose.ui)
+            implementation(libs.compose.preview)
+            implementation(libs.compose.components.resources)
             implementation(libs.adaptive)
             implementation(libs.androidx.lifecycle.viewmodelCompose)
             implementation(libs.androidx.lifecycle.runtimeCompose)
@@ -101,7 +96,7 @@ kotlin {
             implementation(libs.filekit.coil)
         }
         androidMain.dependencies {
-            implementation(compose.preview)
+            implementation(libs.compose.preview)
             implementation(libs.androidx.activity.compose)
             implementation(libs.ktor.client.android)
 
@@ -135,7 +130,6 @@ kotlin {
     }
 
     sourceSets {
-        // ...
         getByName("iosX64Main") {
             kotlin.srcDir("build/generated/ksp/iosX64/iosX64Main/kotlin")
         }
@@ -148,101 +142,10 @@ kotlin {
     }
 }
 
-val projectPackageName = libs.versions.project.packageName.get()
-// load app signing configurations from local.properties file
-val localProperties = getLocalProperties(rootProject)
-
-android {
-    namespace = projectPackageName
-    compileSdk = libs.versions.android.compileSdk.get().toInt()
-
-    defaultConfig {
-        applicationId = projectPackageName
-        minSdk = libs.versions.android.minSdk.get().toInt()
-        targetSdk = libs.versions.android.targetSdk.get().toInt()
-        versionCode = libs.versions.project.versionCode.get().toInt()
-        versionName = libs.versions.project.versionName.get()
-    }
-    packaging {
-        resources {
-            excludes += "/META-INF/{AL2.0,LGPL2.1}"
-        }
-    }
-
-    signingConfigs {
-
-        if (localProperties.hasDebugStoreConfig()) {
-            getByName("debug") {
-                keyAlias = localProperties["DEBUG_KEY_ALIAS"] as? String
-                keyPassword = localProperties["DEBUG_KEY_PASSWORD"] as? String
-                storeFile = file(localProperties["DEBUG_STORE_FILE"] as String)
-                storePassword = localProperties["DEBUG_STORE_PASSWORD"] as? String
-            }
-        }
-        if (localProperties.hasReleaseStoreConfig()) {
-            create("release") {
-                keyAlias = localProperties["RELEASE_KEY_ALIAS"] as? String
-                keyPassword = localProperties["RELEASE_KEY_PASSWORD"] as? String
-                storeFile = file(localProperties["RELEASE_STORE_FILE"] as String)
-                storePassword = localProperties["RELEASE_STORE_PASSWORD"] as? String
-            }
-        }
-    }
-
-
-    buildTypes {
-        debug {
-            signingConfig = signingConfigs.getByName("debug")
-        }
-        getByName("release") {
-            isMinifyEnabled = true
-            isShrinkResources = true
-            isDebuggable = false
-            proguardFiles(
-                getDefaultProguardFile("proguard-android-optimize.txt"),
-                "proguard-rules.pro"
-            )
-            signingConfig = when {
-                localProperties.hasReleaseStoreConfig() -> signingConfigs.getByName("release")
-                else -> signingConfigs.getByName("debug")
-            }
-        }
-    }
-
-    flavorDimensions += "services"
-    productFlavors {
-        //a build flavor with google analytics & crashlytics dependencies
-        create("gms") {
-            isDefault = true
-            dimension = "services"
-        }
-        //a build flavor free of analytics dependencies
-        create("default") {
-            dimension = "services"
-            dependenciesInfo {
-                // Disables dependency metadata when building APKs.
-                includeInApk = false
-                // Disables dependency metadata when building Android App Bundles.
-                includeInBundle = false
-            }
-        }
-    }
-
-    dependencies {
-        "gmsImplementation"(platform(libs.firebase.bom))
-        "gmsImplementation"(libs.firebase.crashlytics.ktx)
-        "gmsImplementation"(libs.firebase.analytics.ktx)
-    }
-
-    compileOptions {
-        sourceCompatibility = JavaVersion.VERSION_11
-        targetCompatibility = JavaVersion.VERSION_11
-    }
-}
 
 compose.desktop {
     application {
-        mainClass = "$projectPackageName.MainKt"
+        mainClass = "${libs.versions.project.packageName.get()}.MainKt"
 
         nativeDistributions {
             targetFormats(TargetFormat.Exe, TargetFormat.Msi, TargetFormat.Dmg, TargetFormat.Deb)
@@ -277,38 +180,16 @@ compose.desktop {
 }
 
 dependencies {
-    debugImplementation(compose.uiTooling)
+    androidRuntimeClasspath(libs.compose.ui.tooling)
 }
 
 // KSP Tasks
 dependencies {
     add("kspCommonMainMetadata", libs.koin.ksp.compiler)
     add("kspAndroid", libs.koin.ksp.compiler)
-    add("kspIosX64", libs.koin.ksp.compiler) // Example for iosX64
+    add("kspIosX64", libs.koin.ksp.compiler)
     add("kspIosArm64", libs.koin.ksp.compiler)
     add("kspIosSimulatorArm64", libs.koin.ksp.compiler)
-}
-
-afterEvaluate {
-    tasks.matching { it.name == "kspDefaultDebugKotlinAndroid" }.configureEach {
-        dependsOn(
-            "generateResourceAccessorsForAndroidDefaultDebug",
-            "generateResourceAccessorsForAndroidMain",
-            "generateActualResourceCollectorsForAndroidMain",
-            "generateResourceAccessorsForAndroidDefault",
-            "generateResourceAccessorsForAndroidDebug"
-        )
-    }
-
-    tasks.named("kspGmsReleaseKotlinAndroid") {
-        dependsOn(
-            "generateResourceAccessorsForAndroidGmsRelease",
-            "generateResourceAccessorsForAndroidMain",
-            "generateActualResourceCollectorsForAndroidMain",
-            "generateResourceAccessorsForAndroidGms",
-            "generateResourceAccessorsForAndroidRelease"
-        )
-    }
 }
 
 // Trigger Common Metadata Generation from Native tasks
@@ -316,21 +197,3 @@ tasks.matching { it.name.startsWith("ksp") && it.name != "kspCommonMainKotlinMet
     .configureEach {
         dependsOn("kspCommonMainKotlinMetadata")
     }
-
-/**
- * Loads properties from the `local.properties` file in the project root.
- *
- * @param root The project root.
- * @return Properties loaded from `local.properties`
- */
-fun getLocalProperties(root: Project): Properties {
-    val localProperties = Properties()
-    val localPropertiesFile = root.file("local.properties")
-    if (localPropertiesFile.exists()) {
-        localProperties.load(localPropertiesFile.inputStream())
-    }
-    return localProperties
-}
-
-fun Properties.hasDebugStoreConfig() = this.containsKey("DEBUG_STORE_FILE")
-fun Properties.hasReleaseStoreConfig() = this.containsKey("RELEASE_STORE_FILE")
