+++ composeApp/src/androidMain/kotlin/com/ma7moud3ly/quran/MainActivity.kt
@@ -58,5 +58,3 @@ private fun ConfigureSystemBars(viewModel: SettingsViewModel) {
         }
     }
 }
-
-
