+++ composeApp/src/commonMain/kotlin/com/ma7moud3ly/quran/features/home/index/IndexContent.kt
@@ -28,7 +28,7 @@ import com.ma7moud3ly.quran.ui.MySurface
 import com.ma7moud3ly.quran.ui.MySurfaceRow
 import com.ma7moud3ly.quran.ui.suraNameFontFamily
 import org.jetbrains.compose.resources.stringResource
-import org.jetbrains.compose.ui.tooling.preview.Preview
+import androidx.compose.ui.tooling.preview.Preview
 import quran.composeapp.generated.resources.Res
 import quran.composeapp.generated.resources.audio_file
 import quran.composeapp.generated.resources.index_madina
