+++ composeApp/src/commonMain/kotlin/com/ma7moud3ly/quran/features/home/reciters/RecitersContent.kt
@@ -31,7 +31,7 @@ import com.ma7moud3ly.quran.ui.AppTheme
 import com.ma7moud3ly.quran.ui.MySurface
 import com.ma7moud3ly.quran.ui.RoundButton
 import org.jetbrains.compose.resources.painterResource
-import org.jetbrains.compose.ui.tooling.preview.Preview
+import androidx.compose.ui.tooling.preview.Preview
 import quran.composeapp.generated.resources.Res
 import quran.composeapp.generated.resources.close
 import quran.composeapp.generated.resources.icon
