+++ composeApp/src/commonMain/kotlin/com/ma7moud3ly/quran/features/reading/ReadingContent.kt
@@ -56,7 +56,7 @@ import kotlinx.coroutines.launch
 import org.jetbrains.compose.resources.DrawableResource
 import org.jetbrains.compose.resources.painterResource
 import org.jetbrains.compose.resources.stringResource
-import org.jetbrains.compose.ui.tooling.preview.Preview
+import androidx.compose.ui.tooling.preview.Preview
 import quran.composeapp.generated.resources.Res
 import quran.composeapp.generated.resources.audio_file
 import quran.composeapp.generated.resources.copied
