+++ composeApp/src/commonMain/kotlin/com/ma7moud3ly/quran/features/reading/VersesScrollbar.kt
@@ -31,7 +31,7 @@ import com.ma7moud3ly.quran.managers.VersesManager
 import com.ma7moud3ly.quran.model.testVersesManager
 import com.ma7moud3ly.quran.ui.AppTheme
 import com.ma7moud3ly.quran.ui.hafsSmartFamily
-import org.jetbrains.compose.ui.tooling.preview.Preview
+import androidx.compose.ui.tooling.preview.Preview
 
 @Preview
 @Composable
