+++ composeApp/src/commonMain/kotlin/com/ma7moud3ly/quran/features/reading/modes/MultiVerses.kt
@@ -44,7 +44,7 @@ import com.ma7moud3ly.quran.ui.hafsSmartFamily
 import kotlinx.coroutines.FlowPreview
 import kotlinx.coroutines.launch
 import org.jetbrains.compose.resources.Font
-import org.jetbrains.compose.ui.tooling.preview.Preview
+import androidx.compose.ui.tooling.preview.Preview
 import kotlin.math.abs
