+++ composeApp/src/commonMain/kotlin/com/ma7moud3ly/quran/features/recitation/config/RecitationConfigContent.kt
@@ -85,7 +85,7 @@ import com.ma7moud3ly.quran.ui.hafsSmartFamily
 import com.ma7moud3ly.quran.ui.isCompactDevice
 import org.jetbrains.compose.resources.StringResource
 import org.jetbrains.compose.resources.stringResource
-import org.jetbrains.compose.ui.tooling.preview.Preview
+import androidx.compose.ui.tooling.preview.Preview
 import quran.composeapp.generated.resources.Res
 import quran.composeapp.generated.resources.downloaded
 import quran.composeapp.generated.resources.recite
