+++ composeApp/src/commonMain/kotlin/com/ma7moud3ly/quran/features/recitation/config/verses/SelectVerseNumber.kt
@@ -40,7 +40,7 @@ import com.ma7moud3ly.quran.ui.MyMiniDialog
 import com.ma7moud3ly.quran.ui.MySurface
 import com.ma7moud3ly.quran.ui.hafsSmartFamily
 import org.jetbrains.compose.resources.stringResource
-import org.jetbrains.compose.ui.tooling.preview.Preview
+import androidx.compose.ui.tooling.preview.Preview
 import quran.composeapp.generated.resources.Res
 import quran.composeapp.generated.resources.ok
 import quran.composeapp.generated.resources.recite_select_begin
