+++ composeApp/src/commonMain/kotlin/com/ma7moud3ly/quran/features/recitation/download/DownloadContent.kt
@@ -34,7 +34,7 @@ import kotlinx.coroutines.delay
 import kotlinx.coroutines.launch
 import org.jetbrains.compose.resources.StringResource
 import org.jetbrains.compose.resources.stringResource
-import org.jetbrains.compose.ui.tooling.preview.Preview
+import androidx.compose.ui.tooling.preview.Preview
 import quran.composeapp.generated.resources.Res
 import quran.composeapp.generated.resources.download_audio_quality
 import quran.composeapp.generated.resources.download_cancel
