+++ composeApp/src/commonMain/kotlin/com/ma7moud3ly/quran/features/recitation/play/PlaybackModeNormal.kt
@@ -54,7 +54,7 @@ import kotlinx.coroutines.FlowPreview
 import kotlinx.coroutines.launch
 import org.jetbrains.compose.resources.painterResource
 import org.jetbrains.compose.resources.stringResource
-import org.jetbrains.compose.ui.tooling.preview.Preview
+import androidx.compose.ui.tooling.preview.Preview
 import quran.composeapp.generated.resources.Res
 import quran.composeapp.generated.resources.back
 import quran.composeapp.generated.resources.copied
