+++ composeApp/src/commonMain/kotlin/com/ma7moud3ly/quran/features/search/SearchContent.kt
@@ -81,7 +81,7 @@ import kotlinx.coroutines.flow.Flow
 import kotlinx.coroutines.flow.flowOf
 import org.jetbrains.compose.resources.StringResource
 import org.jetbrains.compose.resources.stringResource
-import org.jetbrains.compose.ui.tooling.preview.Preview
+import androidx.compose.ui.tooling.preview.Preview
 import quran.composeapp.generated.resources.Res
 import quran.composeapp.generated.resources.audio_file
 import quran.composeapp.generated.resources.search_chapters
