+++ composeApp/src/commonMain/kotlin/com/ma7moud3ly/quran/features/search/chapter/PickChapterContent.kt
@@ -32,7 +32,7 @@ import com.ma7moud3ly.quran.ui.MyDialog
 import com.ma7moud3ly.quran.ui.MySurfaceRow
 import com.ma7moud3ly.quran.ui.suraNameFontFamily
 import org.jetbrains.compose.resources.stringResource
-import org.jetbrains.compose.ui.tooling.preview.Preview
+import androidx.compose.ui.tooling.preview.Preview
 import quran.composeapp.generated.resources.Res
 import quran.composeapp.generated.resources.recite_chapter_select
 import quran.composeapp.generated.resources.search
