+++ composeApp/src/commonMain/kotlin/com/ma7moud3ly/quran/features/search/reciter/PickReciterContent.kt
@@ -40,7 +40,7 @@ import com.ma7moud3ly.quran.ui.MyButton
 import com.ma7moud3ly.quran.ui.MyDialog
 import com.ma7moud3ly.quran.ui.MySurfaceRow
 import org.jetbrains.compose.resources.stringResource
-import org.jetbrains.compose.ui.tooling.preview.Preview
+import androidx.compose.ui.tooling.preview.Preview
 import quran.composeapp.generated.resources.Res
 import quran.composeapp.generated.resources.recite_reciter_select
 import quran.composeapp.generated.resources.reciter_add_all
