+++ composeApp/src/commonMain/kotlin/com/ma7moud3ly/quran/features/settings/SettingsContent.kt
@@ -59,7 +59,7 @@ import org.jetbrains.compose.resources.Font
 import org.jetbrains.compose.resources.StringResource
 import org.jetbrains.compose.resources.painterResource
 import org.jetbrains.compose.resources.stringResource
-import org.jetbrains.compose.ui.tooling.preview.Preview
+import androidx.compose.ui.tooling.preview.Preview
 import quran.composeapp.generated.resources.Res
 import quran.composeapp.generated.resources.continues_verses
 import quran.composeapp.generated.resources.multiple_verses
