+++ gradle/libs.versions.toml
@@ -1,5 +1,6 @@
 [versions]
 #project
+foundation = "1.10.1"
 project-packageName = "com.ma7moud3ly.quran"
 project-vendor = "github.com/Ma7moud3ly"
 project-versionCode = "5"
@@ -8,32 +9,33 @@ android-compileSdk = "36"
 android-minSdk = "24"
 android-targetSdk = "36"
 #plugins
-agp = "8.11.1" # https://www.jetbrains.com/help/kotlin-multiplatform-dev/multiplatform-compatibility-guide.html
-kotlin = "2.2.20"
-ksp = "2.2.20-2.0.3"
-composeHotReload = "1.0.0-beta08"
-composeMultiplatform = "1.9.0"
-kotlinSerialization = "2.2.20"
+agp = "9.0.1" # https://www.jetbrains.com/help/kotlin-multiplatform-dev/multiplatform-compatibility-guide.html
+kotlin = "2.3.10"
+ksp = "2.3.4"
+composeMultiplatform = "1.10.1"
+kotlinSerialization = "2.3.10"
 #common
-ktorClientCore = "3.3.0"
-media3Exoplayer = "1.8.0"
-okio = "3.16.0"
-serialization-json = "1.9.0"
+compose-material3 = "1.9.0"
+compose-material-icons = "1.7.3"
+ktorClientCore = "3.4.0"
+media3Exoplayer = "1.9.2"
+okio = "3.16.4"
+serialization-json = "1.10.0"
 kotlinxDatetime = "0.7.1"
-androidx-lifecycle = "2.9.4"
-androidx-navigation = "2.9.0"
+androidx-lifecycle = "2.9.6"
+androidx-navigation = "2.9.2"
 koinCore = "4.1.1"
 coilCompose = "3.3.0"
-koinAnnotations = "2.1.0"
-adaptive = "1.1.2"
+koinAnnotations = "2.3.1"
+adaptive = "1.2.0"
 multiplatformSettings = "1.3.0"
-filekit = "0.11.0"
+filekit = "0.12.0"
 
 #android
-androidx-activity = "1.11.0"
+androidx-activity = "1.12.4"
 kotlinx-coroutines = "1.10.2"
-firebaseBom = "34.3.0"
-googleServices = "4.4.3"
+firebaseBom = "34.9.0"
+googleServices = "4.4.4"
 firebaseCrashlytics = "3.0.6"
 
 #jvm
@@ -54,6 +56,15 @@ koin-compose = { module = "io.insert-koin:koin-compose", version.ref = "koinCore
 koin-compose-viewmodel = { module = "io.insert-koin:koin-compose-viewmodel ", version.ref = "koinCore" }
 koin-annotations = { module = "io.insert-koin:koin-annotations", version.ref = "koinAnnotations" }
 koin-ksp-compiler = { module = "io.insert-koin:koin-ksp-compiler", version.ref = "koinAnnotations" }
+
+compose-runtime = { module = "org.jetbrains.compose.runtime:runtime", version.ref = "composeMultiplatform" }
+compose-foundation = { module = "org.jetbrains.compose.foundation:foundation", version.ref = "composeMultiplatform" }
+compose-material3 = { module = "org.jetbrains.compose.material3:material3", version.ref = "compose-material3" }
+compose-materialIconsExtended = { module = "org.jetbrains.compose.material:material-icons-extended", version.ref = "compose-material-icons" }
+compose-ui = { module = "org.jetbrains.compose.ui:ui", version.ref = "composeMultiplatform" }
+compose-ui-tooling = { module = "org.jetbrains.compose.ui:ui-tooling", version.ref = "composeMultiplatform" }
+compose-preview = { module = "org.jetbrains.compose.ui:ui-tooling-preview", version.ref = "composeMultiplatform" }
+compose-components-resources = { module = "org.jetbrains.compose.components:components-resources", version.ref = "composeMultiplatform" }
 
 settings-noArg = { module = "com.russhwolf:multiplatform-settings-no-arg", version.ref = "multiplatformSettings" }
 settings-coroutines = { module = "com.russhwolf:multiplatform-settings-coroutines", version.ref = "multiplatformSettings" }
@@ -95,7 +106,7 @@ ktor-client-js = { module = "io.ktor:ktor-client-js", version.ref = "ktorClientC
 [plugins]
 androidApplication = { id = "com.android.application", version.ref = "agp" }
 androidLibrary = { id = "com.android.library", version.ref = "agp" }
-composeHotReload = { id = "org.jetbrains.compose.hot-reload", version.ref = "composeHotReload" }
+androidKmpLibrary = { id = "com.android.kotlin.multiplatform.library", version.ref = "agp" }
 composeMultiplatform = { id = "org.jetbrains.compose", version.ref = "composeMultiplatform" }
 composeCompiler = { id = "org.jetbrains.kotlin.plugin.compose", version.ref = "kotlin" }
 kotlinMultiplatform = { id = "org.jetbrains.kotlin.multiplatform", version.ref = "kotlin" }
